+++ R/CohortDefinition.R
@@ -128,7 +128,10 @@ insertCohortDefinitionSetInPackage <- function(fileName = "inst/settings/Cohorts
   if (insertCohortCreationR && !grepl("inst", fileName))
     stop("When generating R code, the input CSV file must be in the inst folder.")
 
-  cohortsToCreate <- read.csv(fileName)
+  cohortsToCreate <- readr::read_csv(file = fileName,
+                                     col_types = readr::cols(),
+                                     guess_max = 1e+07,
+                                     locale = readr::locale(encoding = "UTF-8"))
 
   # Inserting cohort JSON and SQL
   for (i in 1:nrow(cohortsToCreate)) {
@@ -151,10 +154,10 @@ insertCohortDefinitionSetInPackage <- function(fileName = "inst/settings/Cohorts
   if (generateStats) {
     writeLines("Storing information on inclusion rules")
     rules <- .getCohortInclusionRules(jsonFolder)
-    rules <- merge(rules, data.frame(cohortId = cohortsToCreate$cohortId,
-                                     cohortName = cohortsToCreate$name))
+    rules <- dplyr::inner_join(rules, tidyr::tibble(cohortId = cohortsToCreate$cohortId,
+                                                    cohortName = cohortsToCreate$name))
     csvFileName <- file.path(jsonFolder, "InclusionRules.csv")
-    write.csv(rules, csvFileName, row.names = FALSE)
+    readr::write_csv(x = rules, path = csvFileName)
     writeLines(paste("- Created CSV file:", csvFileName))
   }
 
@@ -182,18 +185,18 @@ insertCohortDefinitionSetInPackage <- function(fileName = "inst/settings/Cohorts
 }
 
 .getCohortInclusionRules <- function(jsonFolder) {
-  rules <- data.frame()
+  rules <- tidyr::tibble()
   for (file in list.files(path = jsonFolder, pattern = ".*\\.json")) {
     writeLines(paste("Parsing", file, "for inclusion rules"))
-    definition <- jsonlite::read_json(file.path(jsonFolder, file))
+    definition <- RJSONIO::fromJSON(file.path(jsonFolder, file))
     if (!is.null(definition$InclusionRules)) {
       nrOfRules <- length(definition$InclusionRules)
       if (nrOfRules > 0) {
         cohortName <- sub(".json", "", file)
         for (i in 1:nrOfRules) {
-          rules <- rbind(rules, data.frame(cohortName = cohortName,
-                                           ruleSequence = i - 1,
-                                           ruleName = definition$InclusionRules[[i]]$name))
+          rules <- dplyr::bind_rows(rules, tidyr::tibble(cohortName = cohortName,
+                                                         ruleSequence = i - 1,
+                                                         ruleName = definition$InclusionRules[[i]]$name))
         }
       }
     }