+++ NEWS.md
@@ -1,6 +1,15 @@
 News
 ================
 
+# ROhdsiWebApi 1.1.2
+
+### Changes:
+1. Replace read.csv with readr::read_csv, write.csv with readr::write_csv to ensure type stability and use of tibble.
+2. Replaced jsonlite with RJSONIO to address issue with special characters "Sjögren"
+3. Replaced used of data.frame(), tidyr::tibble()
+
+### Bug fixes
+
 
 # ROhdsiWebApi 1.1.1